Revert "Clean up int/double castings"
This reverts commit d2c1f2b2bb2298fdb76a707c31eec5900a80335b.

+++ app/src/org/odk/collect/android/views/ResizingImageView.java
@@ -156,15 +156,15 @@ public class ResizingImageView extends ImageView {
         float dHeight = dipToPixels(getContext(), drawable.getIntrinsicHeight());
         float ratio = (dWidth) / dHeight;
 
-        double width = Math.min(Math.max(dWidth, getSuggestedMinimumWidth()), maxWidth);
-        double height = width / ratio;
+        double width = (int) Math.min(Math.max(dWidth, getSuggestedMinimumWidth()), maxWidth);
+        double height = (int) (width / ratio);
 
         height = Math.min(Math.max(height, getSuggestedMinimumHeight()), maxHeight);
-        width = height * ratio;
+        width = (int) (height * ratio);
 
         if (width > maxWidth) {
             width = maxWidth;
-            height = width / ratio;
+            height = (int) (width / ratio);
         }
 
         return new Pair<Integer, Integer>(new Double(width).intValue(), new Double(height).intValue());
